+++ modules/terraform-aci-nd-ra-prefix-policy/outputs.tf
@@ -1,0 +1,9 @@
+output "dn" {
+  value       = aci_rest_managed.ndPfxPol.id
+  description = "Distinguished name of `ndPfxPol` object."
+}
+
+output "name" {
+  value       = aci_rest_managed.ndPfxPol.content.name
+  description = "ND RA prefix policy name."
+}